requirements.txt
playwright==1.43.0
google-api-python-client
google-auth

